==> notebooks/02-joining_main_data.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

%matplotlib inline

### Loading global information

birth_ratio = pd.read_csv("../data/gender-ratio-at-birth.csv")
ratio_at_5 = pd.read_csv("../data/sex-ratio-at-5-years.csv")
ratio_age = pd.read_csv("../data/sex-ratio-by-age.csv")
chao_ratio = pd.read_csv("../data/sex-ratio-at-birth-chao.csv")
female_pop = pd.read_csv("../data/female-population-rate.csv")
infant_mort = pd.read_csv("../data/infant-death-cause-global.csv")

birth_ratio.head()

birth_ratio.tail()

birth_ratio.loc[birth_ratio["Country"] == "World"]

chao_ratio.head()

chao_ratio.tail()

ratio_at_5.head()

ratio_age.head()

female_pop.head()

female_pop.tail()

infant_mort.head()

### Merging following data:

- female population
- sex ratio at birth (Chao et al)
- sex ratio at 5
- sex ratio per age

table should look like:

| Country | Code | Year | Female pop rate | Sex ratio at birth | 5 years old | ... |

female_pop.shape

birth_ratio.shape

chao_ratio.shape

ratio_at_5.shape

ratio_age.shape

world = birth_ratio.loc[birth_ratio["Country"] == "World"]

chao_ratio = pd.concat([chao_ratio, world], ignore_index = True)

merged = chao_ratio.merge(female_pop, on = ["Country", "Code", "Year"], how = 'left')

merged.shape

merged.head()

merged = merged.merge(ratio_at_5, on = ["Country", "Code", "Year"], how = 'left')

merged = merged.merge(ratio_age, on = ["Country", "Code", "Year"], how = 'left')

merged

---

### China and India information

infant_death_india = pd.read_csv("../data/infant-death-cause-india.csv")
missing_female_birth = pd.read_csv("../data/missing-female-birth-ch-in.csv")
missing_women = pd.read_csv("../data/missing-women-ch-in.csv")
birth_order = pd.read_csv("../data/sex-ratio-by-bith-order-ch-kr.csv")

infant_death_india.head()

infant_mort

#### Join India infant mortality rate with global information

infant_mort.shape

infant_death_india.shape

infant_mortality = infant_death_india.merge(infant_mort, on = ["Reason", "Year"], how = 'left')

infant_mortality

---

#### Missing women

missing_female_birth.head()

missing_female_birth.shape

missing_women.head()

missing_women.shape

missing_women.loc[missing_women["Country"] == "Rest of World"]

missing_women = missing_women.merge(missing_female_birth, on = ["Country", "Code", "Year"], how = 'left')

missing_women

missing_women.loc[missing_women["Country"] == "Rest of World"]

missing_women.to_csv("../data/missing-women-merged.csv", index = False)

### Joining missing women with main data

merged = merged.merge(missing_women, on = ["Country", "Code", "Year"], how = 'outer')

merged.head()

merged.loc[merged["Country"] == "Rest of World"]

merged.to_csv("../data/main-merged.csv", index = False)
