==> notebooks/03-reading_and_cleaning_new_data.ipynb <==
import pandas as pd
import numpy as np

country_names = ['Afghanistan','Aland Islands','Albania','Algeria','American Samoa','Andorra','Angola','Anguilla',
                 'Antarctica','Antigua and Barbuda','Argentina','Armenia','Aruba','Australia','Austria','Azerbaijan',
                 'Bahamas','Bahrain','Bangladesh','Barbados','Belarus','Belgium','Belize','Benin','Bermuda','Bhutan',
                 'Bolivia','Bonaire, Saint Eustatius and Saba ','Bosnia and Herzegovina','Botswana','Bouvet Island',
                 'Brazil','British Indian Ocean Territory','British Virgin Islands','Brunei','Bulgaria','Burkina Faso',
                 'Burundi','Cambodia','Cameroon','Canada','Cape Verde','Cayman Islands','Central African Republic','Chad',
                 'Chile','China','Christmas Island','Cocos Islands','Colombia','Comoros','Cook Islands','Costa Rica',
                 'Croatia','Cuba','Curacao','Cyprus','Czech Republic','Democratic Republic of the Congo','Denmark',
                 'Djibouti','Dominica','Dominican Republic','East Timor','Ecuador','Egypt','El Salvador','Equatorial Guinea',
                 'Eritrea','Estonia','Ethiopia','Falkland Islands','Faroe Islands','Fiji','Finland','France','French Guiana',
                 'French Polynesia','French Southern Territories','Gabon','Gambia','Georgia','Germany','Ghana','Gibraltar',
                 'Greece','Greenland','Grenada','Guadeloupe','Guam','Guatemala','Guernsey','Guinea','Guinea-Bissau','Guyana',
                 'Haiti','Heard Island and McDonald Islands','Honduras','Hong Kong','Hungary','Iceland','India','Indonesia',
                 'Iran','Iraq','Ireland','Isle of Man','Israel','Italy','Ivory Coast','Jamaica','Japan','Jersey','Jordan',
                 'Kazakhstan','Kenya','Kiribati','Kosovo','Kuwait','Kyrgyzstan','Laos','Latvia','Lebanon','Lesotho','Liberia',
                 'Libya','Liechtenstein','Lithuania','Luxembourg','Macao','Macedonia','Madagascar','Malawi','Malaysia',
                 'Maldives','Mali','Malta','Marshall Islands','Martinique','Mauritania','Mauritius','Mayotte','Mexico',
                 'Micronesia','Moldova','Monaco','Mongolia','Montenegro','Montserrat','Morocco','Mozambique','Myanmar',
                 'Namibia','Nauru','Nepal','Netherlands','Netherlands Antilles','New Caledonia','New Zealand','Nicaragua',
                 'Niger','Nigeria','Niue','Norfolk Island','North Korea','Northern Mariana Islands','Norway','Oman','Pakistan',
                 'Palau','Palestinian Territory','Panama','Papua New Guinea','Paraguay','Peru','Philippines','Pitcairn','Poland',
                 'Portugal','Puerto Rico','Qatar','Republic of the Congo','Reunion','Romania','Russia','Rwanda','Saint Barthelemy',
                 'Saint Helena','Saint Kitts and Nevis','Saint Lucia','Saint Martin','Saint Pierre and Miquelon','Saint Vincent and the Grenadines',
                 'Samoa','San Marino','Sao Tome and Principe','Saudi Arabia','Senegal','Serbia','Serbia and Montenegro','Seychelles',
                 'Sierra Leone','Singapore','Sint Maarten','Slovakia','Slovenia','Solomon Islands','Somalia','South Africa','South Georgia and the South Sandwich Islands',
                 'South Korea','South Sudan','Spain','Sri Lanka','Sudan','Suriname','Svalbard and Jan Mayen','Swaziland','Sweden',
                 'Switzerland','Syria','Taiwan','Tajikistan','Tanzania','Thailand','Togo','Tokelau','Tonga','Trinidad and Tobago',
                 'Tunisia','Turkey','Turkmenistan','Turks and Caicos Islands','Tuvalu','U.S. Virgin Islands','Uganda','Ukraine',
                 'United Arab Emirates','United Kingdom','United States','United States Minor Outlying Islands','Uruguay','Uzbekistan',
                 'Vanuatu','Vatican','Venezuela','Vietnam','Wallis and Futuna','Western Sahara','Yemen','Zambia','Zimbabwe','World', 'Rest of World']

### Fertility vs mortality

fert_vs_mort = pd.read_csv("../data/original_data/fertility-vs-child-mortality.csv")

fert_vs_mort.head()

fert_vs_mort.drop(columns = "Total population (Gapminder)", inplace = True)

fert_vs_mort.rename(columns = {"Entity" : "Country", "Fertility Rate (Children per woman) (live births per woman)" : "Fertility Rate"}, inplace = True)

fert_vs_mort.dtypes

fert_vs_mort.isna().sum()

fert_vs_mort = fert_vs_mort.loc[fert_vs_mort["Country"].isin(country_names)]

fert_vs_mort.isna().sum()

fert_vs_mort.dropna(inplace = True)

fert_vs_mort["Year"] = pd.to_datetime(fert_vs_mort["Year"], format = "%Y")

fert_vs_mort[["Child mortality rate (%)", "Fertility Rate"]] = fert_vs_mort[["Child mortality rate (%)", "Fertility Rate"]].round(decimals = 2)

fert_vs_mort.head()

fert_vs_mort.to_csv("../data/fertility-vs-mortality.csv", index = False)

----

### Life expectancy

life_exp = pd.read_csv("../data/original_data/life-expectancy-of-women-vs-life-expectancy-of-women.csv")

life_exp.head()

life_exp.isna().sum()

life_exp.shape

Life expenctancy has the same problem with missing value as fertility vs mortality. 
Values missing seems to be from 1800s and has no big impact in this project. 

life_exp.rename(columns = {"Entity" : "Country", "Life expectancy of men (years)" : "Life expectancy men", "Life expectancy of women (years)" : "Life expectancy women"}, inplace = True)

life_exp.drop(columns = "Total population (Gapminder)", inplace = True)

life_exp = life_exp[life_exp["Country"].isin(country_names)]

life_exp.isna().sum()

life_exp.head()

life_exp.dropna(inplace = True)

life_exp

life_exp["Year"] = pd.to_datetime(life_exp["Year"], format = "%Y")

life_exp[["Life expectancy men", "Life expectancy women"]] = life_exp[["Life expectancy men", "Life expectancy women"]].round(decimals = 2)

life_exp.dtypes

life_exp.head()

life_exp.to_csv("../data/life_expectancy.csv", index = False)

----

### GDP

gdp = pd.read_csv("../data/original_data/gdp.csv")

gdp.head()

gdp = gdp[gdp["Country"].isin(country_names)]

gdp["Subject Descriptor"].value_counts()

gdp = gdp.loc[(gdp["Subject Descriptor"] == "Gross domestic product per capita, current prices") | (gdp["Subject Descriptor"] == "Inflation, average consumer prices") | (gdp["Subject Descriptor"] == "Unemployment rate") ]

gdp.loc[gdp["Country"] == "India"]

gdp["Subject Descriptor"].value_counts()

gdp["Units"].value_counts()

gdp.drop(columns = ["Subject Notes", "Scale", "Country/Series-specific Notes"], inplace = True)

gdp.isna().sum()

gdp_new = gdp.loc[(gdp["Units"] == "U.S. dollars") | (gdp["Units"] == "Percent change") | (gdp["Units"] == "Percent of total labor force")]

gdp_new.loc[gdp_new["Country"] == "China"]

gdp_new.columns

gdp_new.head()

#### Goal:

|Country|Code|Year|GDP per capita|Inflation percent| Unemployment rate|
|-------|----|----|--------------|-----------------|------------------|
|Afghan |AFG |2015| 594.513      |  -0.662         |  NaN             |
|Afghan |AFG |2016| 560.601      |   4.384         |  NaN             |
|...    |... |


pd.melt(gdp_new, id_vars = ["Code", "Country"], value_vars = ['1980', '1981',
       '1982', '1983', '1984', '1985', '1986', '1987', '1988', '1989', '1990',
       '1991', '1992', '1993', '1994', '1995', '1996', '1997', '1998', '1999',
       '2000', '2001', '2002', '2003', '2004', '2005', '2006', '2007', '2008',
       '2009', '2010', '2011', '2012', '2013', '2014', '2015', '2016', '2017',
       '2018', '2019', '2020', '2021', '2022', '2023', '2024'], var_name = "Year")

Using melt works, but I'll need to create separate dfs for each `Subject Descriptor`

gdp_new.drop(columns = "Units", inplace = True)

gdp_gdp = gdp_new.loc[gdp_new["Subject Descriptor"] == "Gross domestic product per capita, current prices"]

gdp_infla = gdp_new.loc[gdp_new["Subject Descriptor"] == "Inflation, average consumer prices"]

gdp_unem = gdp_new.loc[gdp_new["Subject Descriptor"] == "Unemployment rate"]

gdp_unem.loc[gdp_unem["Country"] == "India"]

No information for unemployment rate for India.

gdp_pivot = pd.melt(gdp_gdp, 
                    id_vars = ["Code", "Country"], 
                    value_vars = ['1980', '1981',
                                  '1982', '1983', '1984', '1985', '1986', '1987', '1988', '1989', '1990',
                                  '1991', '1992', '1993', '1994', '1995', '1996', '1997', '1998', '1999',
                                  '2000', '2001', '2002', '2003', '2004', '2005', '2006', '2007', '2008',
                                  '2009', '2010', '2011', '2012', '2013', '2014', '2015', '2016', '2017',
                                  '2018', '2019', '2020', '2021', '2022', '2023', '2024'],
                    var_name = "Year", 
                    value_name = "GDP per capita (USD)")

gdp_pivot.shape

infla_pivot = pd.melt(gdp_infla, 
                      id_vars = ["Code", "Country"], 
                      value_vars = ['1980', '1981',
                                    '1982', '1983', '1984', '1985', '1986', '1987', '1988', '1989', '1990',
                                    '1991', '1992', '1993', '1994', '1995', '1996', '1997', '1998', '1999',
                                    '2000', '2001', '2002', '2003', '2004', '2005', '2006', '2007', '2008',
                                    '2009', '2010', '2011', '2012', '2013', '2014', '2015', '2016', '2017',
                                    '2018', '2019', '2020', '2021', '2022', '2023', '2024'],
                      var_name = "Year", 
                      value_name = "Inflation, avg cons. prices (% change)")

infla_pivot.shape

unemp_pivot = pd.melt(gdp_unem, 
                      id_vars = ["Code", "Country"], 
                      value_vars = ['1980', '1981',
                                    '1982', '1983', '1984', '1985', '1986', '1987', '1988', '1989', '1990',
                                    '1991', '1992', '1993', '1994', '1995', '1996', '1997', '1998', '1999',
                                    '2000', '2001', '2002', '2003', '2004', '2005', '2006', '2007', '2008',
                                    '2009', '2010', '2011', '2012', '2013', '2014', '2015', '2016', '2017',
                                    '2018', '2019', '2020', '2021', '2022', '2023', '2024'],
                      var_name = "Year", 
                      value_name = "Unemployment rate")

unemp_pivot.shape

gdp_merged = gdp_pivot.merge(infla_pivot, on = ["Code", "Country", "Year"], how = "left")

gdp_merged = gdp_merged.merge(unemp_pivot, on = ["Code", "Country", "Year"], how = "left")

Now that the dataframe looks like I want, is time to finish the cleaning. 

gdp_merged.dtypes

gdp_merged["Year"] = pd.to_datetime(gdp_merged["Year"], format = "%Y")

gdp_merged.head()

type(gdp_merged["GDP per capita (USD)"])

gdp_merged["GDP per capita (USD)"] = gdp_merged["GDP per capita (USD)"].str.replace(",", "")

gdp_merged["GDP per capita (USD)"] = pd.to_numeric(gdp_merged["GDP per capita (USD)"])

gdp_merged["Inflation, avg cons. prices (% change)"] = gdp_merged["Inflation, avg cons. prices (% change)"].str.replace("--", "").str.replace(",", "")

gdp_merged["Inflation, avg cons. prices (% change)"] = pd.to_numeric(gdp_merged["Inflation, avg cons. prices (% change)"])

gdp_merged["Unemployment rate"] = pd.to_numeric(gdp_merged["Unemployment rate"])

gdp_merged.head()

gdp_merged.dtypes

gdp_merged["Unemployment rate"] = gdp_merged["Unemployment rate"].round(decimals = 2)

gdp_merged["Inflation, avg cons. prices (% change)"] = gdp_merged["Inflation, avg cons. prices (% change)"].round(decimals = 2)

gdp_merged["GDP per capita (USD)"] = gdp_merged["GDP per capita (USD)"].round(decimals = 2)

gdp_merged.head()

gdp_merged.to_csv("../data/economic-information.csv", index = False)
