==> notebooks/04-joining_new_data.ipynb <==
import pandas as pd
import numpy as np

main = pd.read_csv("../data/main-merged.csv")
fert = pd.read_csv("../data/fertility-vs-mortality.csv")
expec = pd.read_csv("../data/life_expectancy.csv")
eco = pd.read_csv("../data/economic-information.csv")

main.head()

main.shape

fert.head()

fert.shape

expec.head()

expec.shape

eco.head()

eco.shape

main = main.merge(fert, on = ["Country", "Code", "Year"], how = "left")

main = main.merge(expec, on = ["Country", "Code", "Year"], how = "left")

main = main.merge(eco, on = ["Country", "Code", "Year"], how = "left")

main.head()

main.columns

main.tail()

main.loc[main["Country"] == "Rest of World"]

main.to_csv("../data/main-merged.csv", index = False)
